requirements.txt
flask>=3.1.0
flask-cors>=6.0.0
openai>=2.29.0
Pillow>=12.1.0
python-dotenv>=1.1.0
